Conditionally load the railtie
